requirements.txt
fastapi
supabase
python-dotenv
imapclient